+++ tests/lint/function_naming/potato.m
@@ -1,0 +1,5 @@
+% (c) Copyright 2020 Florian Schanda
+
+function rv = Kitten
+    rv = 12;
+end